requirements.txt
aioboto3==13.1.1
aiobotocore==2.13.1
aiofiles==24.1.0
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aioitertools==0.11.0
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
atlastk==0.13.3
attrs==24.1.0
boto3==1.34.131
botocore==1.34.131
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
exceptiongroup==1.2.2
fastapi==0.112.0
frozenlist==1.4.1
h11==0.14.0
idna==3.7
iso8601==1.1.0
jmespath==1.0.1
msal==1.30.0
multidict==6.0.5
Office365-REST-Python-Client==2.5.11
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
pypika-tortoise==0.1.6
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tortoise-orm==0.21.5
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.5
wrapt==1.16.0
yarl==1.9.4

